open external profile inside participants tab
